+++ secretflow/ml/nn/sl/attack/torch/__init__.py
@@ -11,13 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from .split_async import PYUSLAsyncTFModel
-from .split_state_async import PYUSLStateAsyncTFModel
-from .pipeline import PYUPipelineTFModel
-
-__all__ = [
-    "PYUSLAsyncTFModel",
-    "PYUSLStateAsyncTFModel",
-    "PYUPipelineTFModel",
-]